Moved User Context from index to _app
